requirements.txt
Django==2.1.7
django-crontab==0.7.1
djangorestframework==3.9.2
pytz==2018.9
